Move getFeaturesByModel method
